an approach to fix project component issue
