+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portfolio",
-  "version": "0.1.0",
+  "version": "5.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
