+++ src/data/projectData.js
@@ -1,36 +1,36 @@
-const projectData =  [
-    {
-        title: "School Website",
-        img: "1",
-        techstack: "ReactJs, TailwindCSS",
-        type: "Web",
-        github: "https://github.com/s2sharpit/school",
-        url: "https://bsseniorsecondaryschool.in/",
-    },
-    {
-        title: "Tic-tac-toe",
-        img: "2",
-        techstack: "NextJs, TypeScript",
-        type: "Web",
-        github: "https://github.com/s2sharpit/tic-tac-toe",
-        url: "https://tic-tac-toe-mu-olive.vercel.app/",
-    },
-    {
-        title: "ARB Bookings",
-        img: "3",
-        techstack: "MERN Stack",
-        type: "Web",
-        github: "https://github.com/s2sharpit/booking_app",
-        url: "https://booking-app-lyart.vercel.app/",
-    },
-    {
-        title: "s2sharpit's admin",
-        img: "4",
-        techstack: "ReactJs, SASS",
-        type: "Web",
-        github: "https://github.com/s2sharpit/react-admin-panel",
-        url: "https://react-admin-panel-three.vercel.app/",
-    },
-]
+const projectData = [
+  {
+    title: "School Website",
+    img: "1",
+    techstack: "ReactJs, TailwindCSS",
+    type: "Web",
+    github: "https://github.com/s2sharpit/school",
+    url: "https://bsseniorsecondaryschool.in/",
+  },
+  {
+    title: "Krishi Bazaar",
+    img: "2",
+    techstack: "React Native",
+    type: "App",
+    github: "https://github.com/s2sharpit/krishi-bazaar",
+    url: "https://github.com/s2sharpit/krishi-bazaar",
+  },
+  {
+    title: "ARB Bookings",
+    img: "3",
+    techstack: "MERN Stack",
+    type: "Web",
+    github: "https://github.com/s2sharpit/booking_app",
+    url: "https://booking-app-lyart.vercel.app/",
+  },
+  {
+    title: "s2sharpit's admin",
+    img: "4",
+    techstack: "ReactJs, SASS",
+    type: "Web",
+    github: "https://github.com/s2sharpit/react-admin-panel",
+    url: "https://react-admin-panel-three.vercel.app/",
+  },
+];
 
 export default projectData;
